moving onCreateOptionsMenu and onOptionsItemSelected methods from FlashCardGeneratorFragment to CreateNewDeckFragment... where I meant to put them in the first place
